requirements.txt
requests==2.18.4
GitPython==2.1.8
